+++ lib/main/Synerise.d.ts
@@ -5,10 +5,13 @@ import { NotificationsModule as Notifications } from './modules/NotificationsMod
 import { InjectorModule as Injector } from './modules/InjectorModule';
 import { PromotionsModule as Promotions } from './modules/PromotionsModule';
 import { ContentModule as Content } from './modules/ContentModule';
-import { Error } from './../classes/types/Error';
+import { IError, Error } from './../classes/types/Error';
 declare class SyneriseInitializer {
     private settings?;
     constructor();
+    static isSyneriseInitialized(): boolean;
+    static invokeOnReadyCallback(): void;
+    static invokeOnErrorCallback(errorObject: IError): void;
     withClientApiKey(clientApiKey: string): this;
     withBaseUrl(baseUrl: string): this;
     withDebugModeEnabled(debugModeEnabled: boolean): this;
@@ -16,13 +19,14 @@ declare class SyneriseInitializer {
     withSettings(settings: ISettingsOptions): this;
     init(): void;
     private setupModules;
+    private setup;
 }
 declare class Synerise {
     constructor();
     static Initializer: () => SyneriseInitializer;
+    static isInitialized(): Boolean;
     static onReady(callback: () => void): void;
     static onError(callback: (error: Error) => void): void;
-    static isInitialized(): boolean;
     static changeClientApiKey(clientApiKey: string): void;
     static get Settings(): Settings;
     static get Client(): Client;
@@ -32,4 +36,4 @@ declare class Synerise {
     static get Promotions(): Promotions;
     static get Content(): Content;
 }
-export { Synerise };
+export { Synerise, SyneriseInitializer };